chore: rename worker fixture to auth
